Move fixture existance check into own tst files
