+++ test/index_test.js
@@ -1,8 +1,7 @@
+var path = require('path');
 var fs = require('fs');
 var tape = require('tape');
-var path = require('path');
 var instructions = require('../index.js');
-var constants = require('./constants');
 
 tape.test('v5 directionFromDegree', function(assert) {
     var v5Instructions = instructions('v5', 'en');
@@ -88,8 +87,6 @@ tape.test('v5 laneDiagram', function(assert) {
 });
 
 tape.test('v5 compile', function(t) {
-    var v5Instructions = instructions('v5', 'en');
-
     t.test('language loading', function(assert) {
         var step = {
             maneuver: {
@@ -108,111 +105,8 @@ tape.test('v5 compile', function(t) {
         assert.end();
     });
 
-    t.test('fixtures exist for every type/modifier combinations', function(assert) {
-        var instructions = require('../instructions').get('en');
-        var basePath = path.join(__dirname, 'fixtures', 'v5');
-
-        function underscorify(input) {
-            return input.replace(/ /g, '_');
-        }
-
-        function checkModifiers(type) {
-            constants.modifiers.forEach(function(modifier) {
-                assert.ok(
-                    fs.existsSync(path.join(basePath, underscorify(type), `${underscorify(modifier)}_default.json`)),
-                    `${type}/${modifier}_default`);
-                assert.ok(
-                    fs.existsSync(path.join(basePath, underscorify(type), `${underscorify(modifier)}_destination.json`)),
-                    `${type}/${modifier}_destination`);
-                assert.ok(
-                    fs.existsSync(path.join(basePath, underscorify(type), `${underscorify(modifier)}_name.json`)),
-                    `${type}/${modifier}_name`);
-            });
-        }
-
-        function checkModifiersNoName(type) {
-            // TODO: Remove this function and replace it complately by checkModifiers
-            constants.modifiers.forEach(function(modifier) {
-                // check normal fixture
-                var p = path.join(basePath, underscorify(type), underscorify(modifier) + '.json');
-
-                assert.ok(fs.existsSync(p), type + '/' + modifier);
-
-                // check no_name fixture if should exist
-                var noNamePath = path.join(basePath, underscorify(type), underscorify(modifier) + '_no_name.json');
-
-                if (instructions.v5[type].default.name ||
-                    instructions.v5[type].default.default.name
-                ) {
-                    assert.ok(fs.existsSync(noNamePath), type + '/' + modifier + '/no name');
-                }
-            });
-        }
-
-        var types = constants.types;
-        types.push('other');
-        types.forEach(function(type) {
-            switch(type) {
-            case 'other':
-                [
-                    'invalid_type',
-                    'way_name_ref',
-                    'way_name_ref_name',
-                    'way_name_ref_destinations',
-                    'way_name_ref_mapbox_hack_1',
-                    'way_name_ref_mapbox_hack_2'
-                ].forEach((f) => {
-                    assert.ok(
-                        fs.existsSync(path.join(basePath, 'other', `${f}.json`)),
-                        `${type}/${f}`);
-                });
-                break;
-            case 'turn':
-                checkModifiers(type);
-                break;
-            case 'rotary':
-                [ 'default', 'exit_1', 'name', 'name_exit' ].forEach((s) => {
-                    assert.ok(
-                        fs.existsSync(path.join(basePath, 'rotary', `${s}_default.json`)),
-                        `${type}/${s}_default`);
-                    assert.ok(
-                        fs.existsSync(path.join(basePath, 'rotary', `${s}_destination.json`)),
-                        `${type}/${s}_destination`);
-                    assert.ok(
-                        fs.existsSync(path.join(basePath, 'rotary', `${s}_name.json`)),
-                        `${type}/${s}_name`);
-                });
-
-                // special fixtures for ordinalization
-                for (i = 2; i <= 11; i++) {
-                    assert.ok(
-                        fs.existsSync(path.join(basePath, 'rotary', `exit_${i}.json`)),
-                        `${type}/exit_${i}_default`);
-                };
-                break;
-            case 'roundabout':
-                [ 'default', 'exit' ].forEach((s) => {
-                    assert.ok(
-                        fs.existsSync(path.join(basePath, 'roundabout', `${s}_default.json`)),
-                        `${type}/${s}_default`);
-                    assert.ok(
-                        fs.existsSync(path.join(basePath, 'roundabout', `${s}_destination.json`)),
-                        `${type}/${s}_destination`);
-                    assert.ok(
-                        fs.existsSync(path.join(basePath, 'roundabout', `${s}_name.json`)),
-                        `${type}/${s}_name`);
-                });
-                break;
-            default:
-                checkModifiersNoName(type);
-                break
-            };
-        });
-
-        assert.end();
-    });
-
     t.test('fixtures match generated instructions', function(assert) {
+        var v5Instructions = instructions('v5', 'en');
         var basePath = path.join(__dirname, 'fixtures', 'v5/');
 
         fs.readdirSync(basePath).forEach(function(type) {
